+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog for NeoFS Node
 - IR now supports container creation requests submitted via new `createV2` contract method (#3670)
 - IR structures containers in the contract iteratively (#3670)
 - SN tries new NEP-11 methods `tokens` and `tokensOf` of Container contract (#3701)
+- SN now listens to NEP-11 `transfer` events of Container contract (#3701)
 
 ### Fixed
 - IR panics at graceful shutdown (#3706)